chore: moved db connection package to lib
