+++ lib/database/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "database",
     srcs = ["database.go"],
-    importpath = "github.com/grigoryevandrey/logistics-app/services/addresses/app/database",
+    importpath = "github.com/grigoryevandrey/logistics-app/lib/database",
     visibility = ["//visibility:public"],
     deps = ["@com_github_lib_pq//:pq"],
 )
