+++ services/addresses/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "github.com/grigoryevandrey/logistics-app/services/addresses",
     visibility = ["//visibility:private"],
     deps = [
+        "//lib/database",
         "//services/addresses/app/config",
-        "//services/addresses/app/database",
         "//services/addresses/app/service",
         "//services/addresses/app/transport",
         "@com_github_spf13_viper//:viper",
