+++ services/addresses/main.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"log"
 
+	"github.com/grigoryevandrey/logistics-app/lib/database"
 	"github.com/grigoryevandrey/logistics-app/services/addresses/app/config"
-	"github.com/grigoryevandrey/logistics-app/services/addresses/app/database"
 	"github.com/grigoryevandrey/logistics-app/services/addresses/app/service"
 	"github.com/grigoryevandrey/logistics-app/services/addresses/app/transport"
 	"github.com/spf13/viper"
