+++ services/vehicles/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "github.com/grigoryevandrey/logistics-app/services/vehicles",
     visibility = ["//visibility:private"],
     deps = [
+        "//lib/database",
         "//services/vehicles/app/config",
-        "//services/vehicles/app/database",
         "//services/vehicles/app/service",
         "//services/vehicles/app/transport",
         "@com_github_spf13_viper//:viper",
