+++ services/vehicles/main.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"log"
 
+	"github.com/grigoryevandrey/logistics-app/lib/database"
 	"github.com/grigoryevandrey/logistics-app/services/vehicles/app/config"
-	"github.com/grigoryevandrey/logistics-app/services/vehicles/app/database"
 	"github.com/grigoryevandrey/logistics-app/services/vehicles/app/service"
 	"github.com/grigoryevandrey/logistics-app/services/vehicles/app/transport"
 	"github.com/spf13/viper"
